Add centroid calculation for picking up ropes/nets.
For hunter, we need a way to pickup ropes/nets after a failed attempt at
catching red salamanders. These updates make it so any click event with
an ID containing the text "rope" or "net" triggers an additional
calculation one which computes the centroid of the tile in which the
rope/net is located. This makes it so that we can click the center of
the tile effectively picking up the items.